requirements.txt
google.generativeai
langchain_google_genai
langchain_community
langchain_core
python-dotenv
streamlit
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
PyPDF2
pydantic